Trial for an alternative implementation of `slide`
Allocate a vector for the rolling window only once
and overwrites its values in each iteration.

Leads to unexpected behaviour in (the unusual)
case the vector the sliding window passed as
argument to .f is assigned during the call. e.g.:
`slide(1:3, identity, 2)` returns
`list(NULL, 2:3, 2:3)` instead of
`list(NULL, 1:2, 2:3)`.